requirements.txt
PyQt5
pyglet
msgpack-python